(#16728) Unify handling of log coloring and prefixing
Previously the logging would omit the Notice prefix for log messages at
the notice level. This is a change from how things had worked in 2.7
where all log messages got a prefix of the log level.

This commit unifies all of the handling of log level colorization and
output by using a data-driven style and by doing so re-adds the Notice
prefix for the notice log level.