Load face and actions correctly for help
This fixes up the face loading in help to deal with errors correctly and
also changes how actions are loaded to use a simple require. In order to
get information back to the user about the status of the face, the load
errors are part of the face and are shown in face specific help when
there are any errors loading action files.