Initialize settings using the external interface
The Puppet::Util::CommandLine code used to use a different path to
initialize the settings from what was exposed to the outside world. This
allowed for us to not actually expose an interface that initialized the
settings in a useful manner.

This changes it so that CommandLine uses the Puppet.initialize_settings
method. In order to do this a large number of assumptions made
throughout the code about Puppet being fully initialized when they are
loaded had to be removed. An obvious example of this is the removal of
the cached_attr on Puppet::Node::Environment because it relied on using
Puppet[] to generate a method at load time.

This also had to remove the restriction that application defaults can
only be initialized once because otherwise the application code would
not be able to set its own settings. There may be bugs lurking in that
related to caching, but the code is so inscrutible we'll have to
continue hacking away at the underbrush to see.