(#11331) Add transformation from pops model to AST
This transformer is intended to be used after parsing to a pops
model. It transforms to a 3.1 AST. The intent is to ensure the exact
same behavior when evaluating. (This is a step on a longer migration
path).

Add Factory#respond_to? since method_missing is implemented.

A method_missing implementation should always also implement
respond_to?

(maint) Fix comment stating that [] is "probably never used".
It is certainly used for testing.

Add AST::BlockExpression and use this instead of ASTArray for statements

An AST::BlockExpression is now used instead of ASTArray to represent
a block of expressions/sequence of expressions when the intent is not
to collect the result of each expression/statement (as ASTArray does).

The grammar is changed to reflect this, and so is merging of classes
that now compose the result instead of appending (modifying).

Name classes with full name when parsing (like 3.1).

It is required that classes are fully named when parsed (i.e.
if nested, they must have the fully expanded name set). This
because they may be lazily loaded. This is required in order to
be able to transform a pops model to 3.1 AST.

(w.i.p) transformer is now wired for lambdas.

This add transformation of pops model to AST for lambda related
expressions. (This is w.i.p, there are no tests yet).