(#11331) Add import support to transformation, fix transforms.
This adds support for importing by instantiating the transformer with
an "importer" (capable of performing an import while transforming).

Several transformations that ended up setting "empty code" where instead
nothing should have been passed now corrected.

Turn on rescue/rethrow in parser support.

This turns on rescue and raise of parse exception. (When developing the
grammar part it is beneficial to get the underlying exception with the
grammar logic location/errors). This hides those details.