(#18764) Modify design of enumerables. Fix doc of all.
Rework each_slice/slice and fix doc of all enumerable  f's

The idea that each/foreach/each_slice should produce the last produced value
from the lambda was bad. This makes it difficult to chain these
calls. These functions now produce the receiver.

Modify the return value of each_slice to the receiver.

Rename each_slice to slice and make block optional to produce array.

It was a bad idea to implement each_slice as the only way to get slices
from an array. The renamed function works as before when given an
optional lambda. When called without a lambda it produces the collected
result of all slices. Thus making it possible to iterate over it.

The alternative was to equip all other iteration methods with slicing
functionality, or to implement a separate slicing function.

Corrects documentation in all enumerable functions.