(#11331) Change QueryExpression to being an Expression instead of UnaryExpression
This because an UnaryExpression has a required expression, and
a QueryExpression may have nil (i.e. query all). This issue was
deteced when writing validation and checking multiplicity.

Modify Model Factory to handle QueryExpression not being Unary.

The factory failed to build query expressions as a consequence of
the refactored superclass of QueryExpression (it was built as
an unary expression earlier).