(#11331) Add first working implementation of Validation to egrammar impl.
This adds validation to the parsing pipeline.
The ParserFactory obtains a Validator configured by a
ValidatorFactory. The ValidatorFactory assembles the various parts
required to check each element, issue diagnostics, handle configuration
of which issues to process, and produce suitable error messages
given context specific information.

The construct is more elaborate than needed for the usecase of "validate
the parsed result for the current version". This because there are
several downstream use-cases (run in "lint mode" where output of mroe
than one error is wanted, both errors and warnings should be output,
parseable output is wanted for integration with other tools, etc.)

Some refactoring is needed as the current configuration is
somewhat more complicated than actually required.

Unit and Integration tests now run green. There are still issues
with some tests when running them in isolation.