(#11331) Make tests run for both current and future parser.
Tests were hardwired to run for only future/eparser. Now all tests
where parser is involved runs for both current/classic and
future/eparser (except when testing methods; where parsing is only
done in order to construct a call; it is the function that is tested,
not the parsing). The collect method testing includes testing the
various ways a call can be made to it, but this is the only place; this
because it is more work to unwire all the others when the final style
for lambdas gets picked.