(#11331) Make it possible to call initial function in interpolation.
It is not possible to have a function call first in interpolation.
"A random number: ${fqdn_random(30}" did not work because the first
bare word is tokenized as a VARIABLE by the lexer.

This changes the lexer to emit a NAME token if the bare word is not
followed by a '('.