(Maint) Re-align requires of pops code
The Pops code had gotten itself into loops and ended up with a lot of
dependencies on what order tests are loaded. This was caused by
everything trying to require just the small part of the pops code that
it felt it wanted. This idea violates Re-use/Release Equivalency
Principle
(http://stackoverflow.com/questions/63142/the-reuse-release-equivalence-principle-rep)
and caused odd test failures.

This reorganizes the requires of the pops code such that anything using
the pops code should only every require 1 or 2 different files:
puppet/pops/api and puppet/pops/impl. These two files then will pull in
all of the components of their portion of pops in the order needed and
will also ensure that the module structures are in place for each
component.