Make type calculator infer PType for types and PType
The infer method did not handle inference of a type (it resulted
in a Ruby type with reference to the type class).

This commit updates the type calculator with an infer of PObjectType.
This means that all types, including the meta-type PType is typed
as PType. Tests added.

(This is all esoteric, but needs to be done correctly).