Treat all Ruby Integer values as PIntegerType
The Puppet Type System will treat all Ruby Integers as PIntegerType
regardless of if they are Fixnum or Bignum since the boundary between
those two depends on the architecture. A fixnum might be limited to
31 bits in which case anything larger than that will be bignum.

A bignum may potentially be larger than 64 bit. Integers that large
are currently not supported.