Add support for Multibind combinators and refactor to make use of them
This adds multibind combinatators as separate classes (instead of
having nested logic inside of the producers). An instance of
a combinator is always used in the refactored implementation.

This commit also adds support for a puppet dsl lambda defined
combinatator, and a combinator that is produced by injection.
(These two remain untested at this point).