Refactor producers
This commit contains a refactoring of producers, from being nested
lambdas holding state on the stack (in true Functional Monad style) to
explicit instances with explicit intance variables.

It started to become very messy to work with features as well as
debugging, and the injector class started to be very cluttered with all
sorts of lambda producing methods nesting each other.

This commit also makes the multibind producers smarter, instead of using
a lambda combinator (with limited power), they now instead support
things like merge/append, priority. (To be properly documented).

This also means that the Combinator part of the model and support will
be deprecated (it is not functional after this commit).