(maint) Prevent error when 'common' is given in the categories.
The category common is predefined and if included in the input
an "attempt to redefine" was raised. Now it ignores the last category if
it is named common. This means that if common is placed at some other
precedence a wanted errors is raised, but never when it is placed last).