Allow running specs from the top-level
By creating a rakefile at the top level that understands the convetion
of the component names we can now work on the components as a whole. The
top level gemfile also pulls in all of the components to allow runing
the system as a whole.

You can't run the rake tasks as "bundle exec rake puppet-cli:spec",
however, because bundler doesn't seem to work well with the directory
changing.