(WIP) Add reinventory command to puppet cert
There is a lot of refactoring around this that I would like to do.
Firstly the #apply() method should be removed from
Puppet::SSL::CertificateAuthority::Interface and just flattened out into
direct calls. Secondly, there is still a test missing for the cert
application (the Interface code is simple enough to not warrant one of
its own). Thirdly, generally making a pass of small tidy ups in the code
in this area would be useful.