(????) Allow dynamic envs to override dir envs
This changes it so that if an environment loader produces an environment
that doesn't have any module directories that exist, then that
environment will be ignored. If, after trying all configured loaders, it
doesn't find a suitable environment, it will then fall back to the
legacy environment system. This allows directory environments to be
lower priority than dynamic and stanza environments.

The fallback to legacy environments is needed because right now there is
an assumption that there is always an environment available.