(PUP-1244) Reimplment yum & rpm version compare
When rpm or yum install a package, puppet later checks to ensure that
the version that was installed was the version that was desired.
However, the version string that the user specifies may not exactly
match the version string as reported by yum/rpm, even if the tool will
consider them to be equivalent. On example of this is requesting to
install puppet version 3.4.1. The package that is installed will then
report the version as 3.4.1-1.el5 and consider the installation a
failure.

  Error: Could not update: Failed to update to version 3.4.1, got
  version 3.4.1-1.el5 instead

This commit replaces the version comparison logic for the yum and rpm
providers to be based on reimplementations of the respective yum and rpm
version comparison logic in those tools.